+++ build_runner/test/daemon/daemon_test.dart
@@ -47,10 +47,10 @@ void main() {
         ],
       ),
       d.dir('test', [
-        d.file('hello_test.dart', '''
-import 'package:test/test.dart';
+        d.file('hello.dart', '''
 main() {
-  test('hello', () {});
+  // Don't actually use test package to speed up tests
+  print('hello');
 }'''),
       ]),
       d.dir('web', [
@@ -220,6 +220,7 @@ main() {
       await d.dir('a', [
         d.dir('web', [
           d.file('main.dart', '''
+// @dart=2.12
 main() {
   print('goodbye world');
 }'''),
@@ -236,7 +237,7 @@ main() {
       var succeededResult = await client.buildResults.first;
       expect(succeededResult.results.first.status, BuildStatus.succeeded);
       var ddcContent = await File(p.join(d.sandbox, 'a', '.dart_tool', 'build',
-              'generated', 'a', 'web', 'main.unsound.ddc.js'))
+              'generated', 'a', 'web', 'main.sound.ddc.js'))
           .readAsString();
       expect(ddcContent, contains('goodbye world'));
     });